+++ ReactUI/src/App.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { ThemeProvider} from "@material-ui/core/styles";
 import { unstable_createMuiStrictModeTheme } from '@material-ui/core/styles';
 
-import Container from './components/container'
+import Container from './imports/ui/pages/home'
 
 const theme = unstable_createMuiStrictModeTheme();
